+++ memcached/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Update `async-memcached` to version `0.4`
+
 ## [0.3.2] - 2024-05-04
 
 - Update `deadpool` dependency to version `0.12`